rename: 옵션 페이지 관련 컴퍼넌트 OptionPage 폴더에 통합 #50
* /layouts/ -> /layouts/OptionPage/
