+++ frontend-dashboard/src/layouts/OptionPage/OptionPageLayout.jsx
@@ -1,9 +1,9 @@
 import { usePageStore } from "@/stores/usePageStore";
 import Body from "@/components/Body";
-import CursorModeOptions from "@/pages/cursorMode/CursorModeOptions";
+import CursorModeOptions from "@/layouts/OptionPage/CursorModeOptions";
 import Footer from "@/components/Footer";
 import Header from "@/components/Header";
-import ScrollModeOptions from "@/pages/scrollMode/ScrollModeOptions";
+import ScrollModeOptions from "@/layouts/OptionPage/ScrollModeOptions";
 
 export default function OptionPageLayout() {
   const optionPage = usePageStore((state) => state.optionPage);
